Allow for 'none' channel/axis/button
When using interaction it may be needed to control
given mapping from outside. Then we should avoid connecting
that mapping to real serial channel, because everything set
by interaction will be overwritten. That's why now the channel
can be set to 0, which does not read data from the serial port.

Furthermore button 0/axis None can be set. This is useful
if a given serial channel is intended to be used by interaction
but we don't want to send its values to vJoy.