VJoy classes reworked and vXbox support
To support more than one Virtual Joystick driver a new class
VJoyCollectionBase has been implemented. Its task is to collect
the list of the joysticks available from the different drivers
and return the appropriate VJoyBase instance when requested.

vXbox is implemented with pinvoke to the vXboxInterface.dll
which calls the ScpVbus driver.